chpaswd integration

+++ live-installer/installer.py
@@ -120,8 +120,9 @@ class InstallerEngine:
         os.system("cp -f /etc/resolv.conf /target/etc/resolv.conf")
 
         kernelversion = subprocess.getoutput("uname -r")
-        os.system(
-            "cp /lib/modules/{0}/vmlinuz /target/boot/vmlinuz-{0}".format(kernelversion))
+        if os.path.exists("/lib/modules/{0}/vmlinuz".format(kernelversion))
+            os.system(
+                "cp /lib/modules/{0}/vmlinuz /target/boot/vmlinuz-{0}".format(kernelversion))
 
         # add new user
         print(" --> Adding new user")
@@ -143,11 +144,20 @@ class InstallerEngine:
             self.do_run_in_chroot(
                 "usermod -aG {} {}".format(group, self.setup.username))
 
-        self.do_run_in_chroot(
-            "echo -e \"{0}\\n{0}\\n\" | passwd {1}".format(self.setup.password1, self.setup.username))
-        if config.main["set_root_password"]:
+        if os.system("which chpasswd &>/dev/null") == 0:
+            fp = open("/target/tmp/.passwd", "w")
+            fp.write(self.setup.username +  ":" + self.setup.password1 + "\n")
+            if config.main["set_root_password"]:
+                fp.write("root:" + self.setup.password1 + "\n")
+            fp.close()
+            self.do_run_in_chroot("cat /tmp/.passwd | chpasswd")
+            os.system("rm -f /target/tmp/.passwd")
+        else:
             self.do_run_in_chroot(
-                "echo -e \"{0}\\n{0}\\n\" | passwd".format(self.setup.password1))
+                "echo -e \"{0}\\n{0}\\n\" | passwd {1}".format(self.setup.password1, self.setup.username))
+            if config.main["set_root_password"]:
+                self.do_run_in_chroot(
+                    "echo -e \"{0}\\n{0}\\n\" | passwd".format(self.setup.password1))
 
         # Set LightDM to show user list by default
         if config.main["list_users_when_auto_login"]:
